fix: change var type

+++ src/main/scala/ru/shemich/App.scala
@@ -2,9 +2,7 @@ package ru.shemich
 
 import org.apache.spark.sql.{Row, SparkSession}
 import org.apache.spark.sql.functions._
-import org.apache.spark.sql.types.{ArrayType, IntegerType, StringType, StructType}
-
-import java.util.Properties
+import org.apache.spark.sql.types.{ArrayType, LongType, StringType, StructType}
 
 object App {
   def main(args: Array[String]): Unit = {
@@ -43,22 +41,22 @@ object App {
     )
 
     val transactionsStructureData = Seq(
-      Row("Lukoil", "Surgutneftegas", 12000, 1),
-      Row("Lukoil", "Gazprom", 31450, 2),
-      Row("Lukoil", "Rosneft", 5050, 3),
-      Row("Rosneft", "Lukoil", 25000, 4),
-      Row("Rosneft", "Surgutneftegas", 500, 5),
-      Row("Sberbank", "Russian Railways", 50, 6),
-      Row("Sberbank", "Red&White", 11000, 7),
-      Row("Russian Railways", "AvtoVAZ", 3330, 8),
-      Row("Gazprom", "Sberbank", 50600, 9),
-      Row("Red&White", "Sberbank", 3500, 10),
-      Row("Chelyabinsk Pipe", "Sberbank", 1200, 11),
-      Row("AvtoVAZ", "Surgutneftegas", 1111, 12),
-      Row("DNS Group", "AvtoVAZ", 7599, 13),
-      Row("Surgutneftegas", "Gazprom", 1231, 14),
-      Row("Surgutneftegas", "Lukoil", 254, 15),
-      Row("AvtoVAZ", "Sberbank", 12, 16),
+      Row("Lukoil", "Surgutneftegas", 12000L, 1L),
+      Row("Lukoil", "Gazprom", 31450L, 2L),
+      Row("Lukoil", "Rosneft", 5050L, 3L),
+      Row("Rosneft", "Lukoil", 25000L, 4L),
+      Row("Rosneft", "Surgutneftegas", 500L, 5L),
+      Row("Sberbank", "Russian Railways", 50L, 6L),
+      Row("Sberbank", "Red&White", 11000L, 7L),
+      Row("Russian Railways", "AvtoVAZ", 3330L, 8L),
+      Row("Gazprom", "Sberbank", 50600L, 9L),
+      Row("Red&White", "Sberbank", 3500L, 10L),
+      Row("Chelyabinsk Pipe", "Sberbank", 1200L, 11L),
+      Row("AvtoVAZ", "Surgutneftegas", 1111L, 12L),
+      Row("DNS Group", "AvtoVAZ", 7599L, 13L),
+      Row("Surgutneftegas", "Gazprom", 1231L, 14L),
+      Row("Surgutneftegas", "Lukoil", 254L, 15L),
+      Row("AvtoVAZ", "Sberbank", 12L, 16L)
     )
     val regionsStructureSchema = new StructType()
       .add("region", StringType)
@@ -67,8 +65,8 @@ object App {
     val transactionsStructureSchema = new StructType()
       .add("company_from", StringType)
       .add("company_to", StringType)
-      .add("amount", IntegerType)
-      .add("transaction_id", IntegerType)
+      .add("amount", LongType)
+      .add("transaction_id", LongType)
 
     val regionsDF = spark.createDataFrame(
       spark.sparkContext.parallelize(regionsStructureData), regionsStructureSchema)
